+++ src/core/sent_packet_metadata.h
@@ -78,7 +78,7 @@ typedef struct QUIC_SENT_FRAME_METADATA {
     //
     uint64_t StreamOffset;
     uint16_t StreamLength;
-    uint16_t Type; // QUIC_FRAME_*
+    uint32_t Type; // QUIC_FRAME_*
     uint8_t Flags; // QUIC_SENT_FRAME_FLAG_*
 
 } QUIC_SENT_FRAME_METADATA;
